fix(reservoir): extract SparseWeights to meet LOC gate
- Extract SparseWeights struct to reservoir_sparse.rs (101 LOC)
- reservoir.rs now 412 LOC (under 500 limit)
- Add CHANGELOG entry for 0.3.4
